pyproject.toml
[project]
name = "guild-schema-manager"
version = "0.1.0"
description = "Discord Server Schema Manager Bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.4.0,<3.0.0",
  "PyYAML>=6.0.1,<7.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.4,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
